requirements_SR.txt
absl-py==0.15.0
addict==2.4.0
astunparse==1.6.3
backcall==0.2.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
clang==5.0
click==8.0.3
configparser==5.0.2
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
docker-pycreds==0.4.0
dominate==2.6.0
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.18
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
h5py==3.1.0
idna==3.3
imageio==2.9.0
importlib-metadata==4.8.1
ipython==6.2.0
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lmdb==1.2.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
networkx==2.5.1
Nib==0.5.7
nibabel==3.2.1
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.4.58
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
promise==2.3
prompt-toolkit==1.0.18
protobuf==3.19.0
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
pyparsing==2.4.7
pyreadline==2.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
sentry-sdk==1.4.3
sh==1.14.2
shortuuid==1.0.1
simplegeneric==0.8.1
six==1.15.0
sklearn==0.0
smmap==5.0.0
subprocess32==3.5.4
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2020.9.3
torch==1.10.1+cu113
torchaudio==0.10.1+cu113
torchfile==0.1.0
torchvision==0.11.2+cu113
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.7
visdom==0.1.8.9
wandb==0.12.6
wcwidth==0.2.5
websocket-client==1.2.3
Werkzeug==2.0.2
wrapt==1.12.1
yapf==0.31.0
yaspin==2.1.0
zipp==3.6.0